First draft implementation of vertical optimization
Background: In LogicalProjects, so far each column got
calculated independently and was assigned simultaneously at the end.
This unfortunately means, that more complex operations such as
"OVER" need to do a lot of shuffling and book-keeping, as they need
to keep track of the original order (otherwise the columns
will not fit together).
This commit changes this behaviour by adding the columns
immediately to the dataframe, allowing to re-sort the dataframe if
needed.